Public pipeline and step API
